Mob Overhaul Commit 1
This is a massive overhaul to the way Mobs are spawned in the game and how the ai is layed out. This commit adds the ability to spawn in variable waves of enemies of different numbers and types as well as functionality for those enemies to take damage and change types.