Playable Commit
This addition adds mobile friendly controls for placing towers as well as well as many quality of life features. Dying now sends you to the title card, you cannot place towers outside the bounds of the ship. Money is now only taken from you when a tower is actually placed.